% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.constant.categorical.R
\name{check.constant.categorical}
\alias{check.constant.categorical}
\title{Check for constant metrics and categorical metrics}
\usage{
check.constant.categorical(dataset, metrics)
}
\arguments{
\item{dataset}{a data frame for data}

\item{metrics}{a characters or a vector of characters for independent variables}
}
\description{
Check for constant metrics and categorical metrics
}
\examples{
Data = loadDefectDataset('groovy-1_5_7','jira')
check.constant.categorical(dataset = Data$data, metrics = Data$indep)
}
\keyword{categorical}
\keyword{constant}
