% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.parallel.R
\name{fit.parallel}
\alias{fit.parallel}
\title{Universal fit function (Parallel)}
\usage{
fit.parallel(data, dep, indep, classifier = "lr",
  classifier.params = list(rf.ntree = 100, rf.mtry = NULL, c5.0.trials =
  40, c5.0.rules = TRUE, c5.0.winnow = FALSE, nb.fL = 0, nb.adjust = 1,
  svm.gamma = NULL, svm.cost = 1), params.tuning = FALSE,
  normalize = "no", rebalance = "no", validation = "boot",
  validation.params = list(cv.k = 10, boot.n = 100),
  prob.threshold = 0.5, repeats = 1, n.cores = 2)
}
\arguments{
\item{data}{a dataframe for input data}

\item{dep}{a character for dependent variable}

\item{indep}{a vector of characters for independent variables}

\item{classifier}{a character for classifier techniques, i.e., lr, rf, c5.0, nb, and svm}

\item{classifier.params}{a list of parameters for an input classifier technique}

\item{params.tuning}{a boolean indicates whether to perform parameters tuning}

\item{normalize}{a character for normalization techniques, i.e., log, scale, center, standardize, and no for non-normalization#'}

\item{rebalance}{a character for a choice of data sampling techniques, i.e., up for upsampling, down for downsampling, and no for no-sampling (default: "NO")}

\item{validation}{a character for a choice of validation techniques, i.e., boot for bootstrap validation technique, cv for cross-validation technique, and no for constructing a model with the whole dataset without model validation (default: "boot")}

\item{validation.params}{a list of parameters for an input validation techniques (default: list(cv.k = 10, boot.n = 100))}

\item{prob.threshold}{a numeric for probability threshold (default: 0.5)}

\item{repeats}{a numeric for number of repetitions (default: 1)}

\item{n.cores}{a numeric for number of cores (default: 2)}
}
\description{
Universal fit function (Parallel)
}
\keyword{fit}
