% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit}
\alias{fit}
\title{Universal fit function}
\usage{
fit(data, dep, indep, classifier = "LR", classifier.params = list(RF.ntree =
  100, C5.0.trials = 40, C5.0.rules = TRUE), rebalance = "NO",
  validation = "BOOT", validation.params = list(CV.k = 10, BOOT.N = 100),
  prob.threshold = 0.5, repeats = 1)
}
\arguments{
\item{data}{a dataframe for input data}

\item{dep}{a character for dependent variable}

\item{indep}{a vector of characters for independent variables}

\item{classifier}{a character for classifier techniques, i.e., LR, RF, C5.0, and NB (default: "LR")}

\item{classifier.params}{a list of parameters for an input classifier technique (default: list(RF.ntree = 100, C5.0.trials = 40, C5.0.rules = TRUE)}

\item{rebalance}{a character for a choice of data sampling techniques, i.e., UP for upsampling, DOWN for downsampling, and NO for no-sampling (default: "NO")}

\item{validation}{a character for a choice of validation techniques, i.e., BOOT for bootstrap validation technique, and CV for cross-validation technique (default: "BOOT")}

\item{validation.params}{a list of parameters for an input validation techniques (default: list(CV.k = 10, BOOT.N = 100))}

\item{prob.threshold}{a numeric for probability threshold (default: 0.5)}

\item{repeats}{a numeric for number of repetitions (default: 1)}
}
\description{
Universal fit function
}
\keyword{fit}
