% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.importance.R
\name{get.importance}
\alias{get.importance}
\title{Get variable importance for Logistic Regression and Random Forest}
\usage{
get.importance(model, normalized = F)
}
\arguments{
\item{model}{a logistic regression or random forest model}

\item{normalized}{a logical to indicate whether to normalize importance scores}
}
\description{
This function makes life simple by automatically interpreting logistic regression and random forest models.
The interpretation techniques for logistic regression are ANOVA Type-I and variants of ANOVA Type-II (i.e., Wald, Chisq, and F).
The interpretation techniques for random forest are Gini and Permutation (i.e., Scaled and Raw) importance.
}
\keyword{interpretation}
