% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.vc.R
\name{get.vc}
\alias{get.vc}
\title{Get VarClus based on the absolute Spearman correlation coefficients between metrics}
\usage{
get.vc(dataset, metrics, similarity = "spearman",
  varclus.threshold = 0.7)
}
\arguments{
\item{dataset}{a data frame for data}

\item{metrics}{a vector of characters or a vector of characters for independent variables}

\item{similarity}{a character for similarity measures (e.g., Spearman rank correlation), default = spearman}

\item{varclus.threshold}{a numeric for correlation coefficient threshold value}
}
\description{
This function makes life simple by providing a VarClus.
}
\keyword{VarClus}
