% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.calculation.R
\name{performance.calculation}
\alias{performance.calculation}
\title{Performance Calculation}
\usage{
performance.calculation(outcome, prob, threshold = 0.5)
}
\arguments{
\item{outcome}{a vector of the outcome labels.}

\item{prob}{a numeric vector of the predicted probability.}

\item{threshold}{a numeric threshold for the predicted probability. A predicted probability of above a threshold is considered a positive prediction.}
}
\description{
This function exports confusion matrix (i.e., true positive, false positive, true negative, and false negative) and calculates 11 performance measures.
The performance measures are AUC, Precision, Recall, False Positive Rate, True Negative Rate, F-measure, Accuracy, MCC, G-mean, G-measure, and Balance.
}
\keyword{performance}
