% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVarClus.R
\name{plotVarClus}
\alias{plotVarClus}
\title{Get VarClus plot based on the absolute Spearman correlation coefficients between metrics}
\usage{
plotVarClus(dataset, metrics, correlation = "spearman",
  correlation.threshold = 0.7)
}
\arguments{
\item{dataset}{a data frame for data}

\item{metrics}{a vector of characters or a vector of characters for independent variables}

\item{correlation}{a character for correlation measures (e.g., Spearman rank correlation), default = spearman}

\item{correlation.threshold}{a numeric for correlation coefficient threshold value}
}
\description{
This function makes life simple by providing a VarClus plot.
}
\examples{
Data = loadDefectDataset('camel-1.0','ck')
plotVarClus(dataset = Data$data, metrics = Data$indep)
}
\keyword{VarClus}
