% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single.fit.R
\name{single.fit}
\alias{single.fit}
\title{A single fit function}
\usage{
single.fit(training.data, testing.data, dep, indep, classifier = "lr",
  classifier.params = list(rf.ntree = 100, c5.0.trials = 40, c5.0.rules =
  TRUE), params.tuning = F, prob.threshold)
}
\arguments{
\item{training.data}{a dataframe for training data}

\item{testing.data}{a dataframe for testing data}

\item{dep}{a character for dependent variable}

\item{indep}{a vector of characters for independent variables}

\item{classifier}{a character for classifier techniques, i.e., lr, rf, c5.0, and nb (default: "lr")}

\item{classifier.params}{a list of parameters for an input classifier technique (default: list(rf.ntree = 100, c5.0.trials = 40, c5.0.rules = TRUE)}

\item{params.tuning}{a boolean indicates whether to perform parameters tuning}
}
\description{
A single fit function
}
\keyword{fit}
